[cxx-interop] Refactor: move synthesis logic elsewhere
